Changes for kpi
